// src/app/models/Orders.ts
export class Orders {
    Id: number;
    CarId: number;
    OrderDate: Date;
    OrderEndDate: Date;
    CreateDate: Date;
    Gender: boolean;
    FirstName: string;
    LastName: string;
    Email: string;
    Phone: string;
    PersonalId: string;
    Comment: string;
    StatusCode: number;
    StatusDescrip: string;
    TakeLocation: string;
    ReceiveLocation: string;
    TransactionId: string;
    Status: string;
    ResultCode: number;
    Amount:number;
    UniqueId: string;
}